fix working vue-toasted for vuestic-admin
